change upload method, extend cloud disk method

+++ serializer/file.go
@@ -6,7 +6,7 @@ import "github.com/ChenMiaoQiu/go-cloud-disk/model"
 type File struct {
 	Uuid     string `json:"file_id"`
 	FileName string `json:"filename"`
-	Size     uint   `json:"size"`
+	Size     int64  `json:"size"`
 }
 
 func BuildFile(file model.File) File {
