+++ serializer/file_store.go
@@ -6,8 +6,8 @@ import "github.com/ChenMiaoQiu/go-cloud-disk/model"
 type FileStore struct {
 	Uuid        string `json:"id"`
 	OwnerId     string `json:"owner"`
-	MaxSize     uint   `json:"maxsize"`
-	CurrentSize uint   `json:"currentsize"`
+	MaxSize     int64  `json:"maxsize"`
+	CurrentSize int64  `json:"currentsize"`
 }
 
 // BuildFileStore return fileStore serializer
